Use GHA backend for caching (#69)
Rather than writing the cache to a file, this uses Buildkit's support to write directly to the GHA backend.

Because the GHA backend is only available in GHA, the Makefile is not affected. Possibly some of the caching in there could be removed.

The docker/build-push-action doesn't need a checkout, so that is removed in build.yaml. It's left in in release.yaml as I think cucumber/action-get-versions may need it.